requirements.txt
python-dotenv==1.0.0
blocklist_aggregator==1.3.0